ai4code/learning/evaluation/tesseract/requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
pandas==0.22.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
Tesseract==0.9
tqdm==4.25.0
